requirements.txt
# PDF Search Engine - Dependencies

# Web interface
streamlit>=1.28.0

# PDF text extraction
PyPDF2>=3.0.0
pdfplumber>=0.10.0

# pdfplumber transitive dependencies (installed automatically but pinned for reproducibility)
Pillow>=9.0.0
pdfminer.six>=20221105

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Note: The following are Python standard library modules (no install needed):
# - sqlite3, logging, pathlib, dataclasses, json, re, hashlib
# - base64, typing, contextlib, time, argparse, subprocess, sys
